backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
termcolor = "^2.3.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"
bs4 = "^0.0.1"
lxml = "^4.9.3"
fastapi-csrf-protect = "^0.3.2"
mysqlclient = "^2.2.0"
sqlalchemy = "^2.0.22"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
python-decouple = "^3.8"
dnspython = "^2.4.2"
alembic = "^1.12.1"
pymysql = "^1.1.0"
cryptography = "^41.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
fastapi-sessions = "^0.3.2"
dependency-injector = "^4.41.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
